% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_stats.R
\name{get_edge_weight_summary}
\alias{get_edge_weight_summary}
\title{Get summary of edge weights}
\usage{
get_edge_weight_summary(graph, n = 10)
}
\arguments{
\item{graph}{\code{data.frame} containing the graph. must have columns from,
to and weight, so each row defines an edge.}

\item{n}{The desired number of quantiles.}
}
\value{
\code{data.frame} containing the graph's edge weights divided into a
given number of quantiles.
}
\description{
Get summary of edge weights
}
