% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_stats.R
\name{get_graph_sizes}
\alias{get_graph_sizes}
\title{Get number and sizes of graph components}
\usage{
get_graph_sizes(graph)
}
\arguments{
\item{graph}{\code{data.frame} containing the graph. must have columns from
and to, so each row defines an edge.}
}
\value{
\code{vector} containing the sizes of all individual graph components
in descending order.
}
\description{
Get number and sizes of graph components
}
\examples{
\dontrun{
n <- 50
from <- as.integer (runif (min = 0, max = 100, n))
to <- as.integer (runif (min = 0, max = 100, n))
weight <- runif (n)
graph <- data.frame (from, to)

get_graph_sizes (graph = graph)
}
}
