% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_graph_statistics}
\alias{get_graph_statistics}
\title{Get graph component statistics}
\usage{
get_graph_statistics(graph)
}
\arguments{
\item{graph}{\code{data.frame} containing the graph. must have columns from,
to and weight.}
}
\value{
\code{vector} containing the sizes of all individual graph components
in descending order.
}
\description{
Get graph component statistics
}
\examples{
\dontrun{
n <- 50
from <- as.integer (runif (min = 0, max = 100, n))
to <- as.integer (runif (min = 0, max = 100, n))
weight <- runif (n)
graph <- data.frame (from, to, weight)

get_graph_statistics (graph = graph)
}
}
