% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_stats.R
\name{get_num_unique_vertices}
\alias{get_num_unique_vertices}
\title{Get number of unique vertices}
\usage{
get_num_unique_vertices(graph)
}
\arguments{
\item{graph}{\code{data.frame} containing the graph. must have columns from
and to, so each row defines an edge.}
}
\value{
The number of unique vertices in the graph.
}
\description{
Get number of unique vertices
}
