% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_graph_report}
\alias{generate_graph_report}
\title{Generate graph report from a template}
\usage{
generate_graph_report(graph, article_name = "graph-report", n = 10)
}
\arguments{
\item{graph}{\code{data.frame} containing the graph. must have columns from,
to and weight, so each row defines an edge.}

\item{article_name}{file name of the article.}

\item{n}{The desired number of quantiles for the edge weight section.}
}
\description{
Generate graph report from a template
}
