% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_graph_statistics}
\alias{get_graph_statistics}
\title{Get graph component statistics}
\usage{
get_graph_statistics(graph)
}
\arguments{
\item{graph}{\code{data.frame} containing the graph. must have columns from,
to and weight.}
}
\value{
\code{data.frame} containing the number and size of graph components.
}
\description{
Get graph component statistics
}
