% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_stats.R
\name{get_node_degrees}
\alias{get_node_degrees}
\title{Get degrees of all nodes in the graph}
\usage{
get_node_degrees(graph)
}
\arguments{
\item{graph}{\code{data.frame} containing the graph. must have columns from,
to and weight, so each row defines an edge.}
}
\value{
\code{data.frame} the ID and frequency of every node in the graph.
}
\description{
Get degrees of all nodes in the graph
}
